ebe1113 Some more fix hopefully
0dab52a Another rounds of warning clean up

+++ ParaViewCore/ServerManager/Core/vtkSMDataTypeDomain.cxx
@@ -122,7 +122,7 @@ vtkSMDataTypeDomain::~vtkSMDataTypeDomain()
 //---------------------------------------------------------------------------
 unsigned int vtkSMDataTypeDomain::GetNumberOfDataTypes()
 {
-  return this->DTInternals->DataTypes.size();
+  return static_cast<unsigned int>(this->DTInternals->DataTypes.size());
 }
 
 //---------------------------------------------------------------------------
