+++ ParaViewCore/ServerManager/Rendering/vtkSMDataDeliveryManager.cxx
@@ -99,7 +99,7 @@ void vtkSMDataDeliveryManager::Deliver(bool interactive)
          << static_cast<int>(use_lod)
          << static_cast<unsigned int>(keys_to_deliver.size())
          << vtkClientServerStream::InsertArray(
-           &keys_to_deliver[0], keys_to_deliver.size())
+           &keys_to_deliver[0], static_cast<int>(keys_to_deliver.size()))
          << vtkClientServerStream::End;
   this->ViewProxy->GetSession()->ExecuteStream(
     this->ViewProxy->GetLocation(), stream, false);
@@ -140,7 +140,7 @@ bool vtkSMDataDeliveryManager::DeliverStreamedPieces()
            << "DeliverStreamedPieces"
            << static_cast<unsigned int>(keys_to_deliver.size())
            << vtkClientServerStream::InsertArray(
-             &keys_to_deliver[0], keys_to_deliver.size())
+             &keys_to_deliver[0], static_cast<int>(keys_to_deliver.size()))
            << vtkClientServerStream::End;
     session->ExecuteStream(this->ViewProxy->GetLocation(), stream, false);
     }
